// frontend/src/app/core/models/auth.models.ts
export interface AuthenticationRequest {
    email: string;
    password: string;
}

export interface RegisterRequest {
    firstname: string;
    lastname: string;
    email: string;
    password: string;
}

export interface AuthenticationResponse {
    access_token: string;
    refresh_token: string;
}